Moved Leaflet components to modules and update Rust Analyzer settings.
